Updated mappings to 1.20.4
Updated the mappings from 1.16 to 1.20.4!

Items are hidden but the packets are not sent when command is executed. You need to reequip your weapon/armor after the command is executed.

Hade to make an ugly solution with ProtocolLib since a dev build is used(ProtocolLib 1.20.4 is not fully released yet)